Update class-gpnf-session.php
The array( self::COOKIE_NAME, $this->_form_id ) was an array of arrays hence was not getting converted into string by implode. Even running a for loop through each element did not work. So to remove the bug, I used the first element of the array and the current time and date to make it a unique cookie.